Adding the ability to get a quick peek at the pat_wave when it is really long.
Examples:

>>>  When reading the This will return the string up to 96 bits.
>>>
>>>  If the length is more than 96 bits the string will append '... ( 4904 more )'
>>>  >>> cs.pat_wave
>>>  '111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111 ... ( 4904 more )'
>>>   >>>
>>>
>>>  **NOTE**: To get pattern waves that are longer than 96 use the get_pat_wave function.
>>>
>>>  example:
>>>   >>> cs.pat_wave = '11001'
>>>   >>> cs.pat_wave
>>>   '11001'
>>>   >>> cs.pat_wave = '1'*97
>>>   >>> cs.pat_wave
>>>   '111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111 ... ( 1 more )'
>>>   >>> cs.get_pat_wave()
>>>   '1111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111'
>>>   >>>

chipshouter.py
==============
added get_pat_length() -> Just calls the get_pat_length() of the com_tools.py
Modified pat_wave reading to check the length and it will append ( X more)

96
--
The reason for 96 is just because of the number of bits we can get in one
packet.

com_tools.py
============
Added PAT_WAVE_LENGTH option to get from the firmware.

Added method to get_pat_length from the firmware.
Added paramater 'short_wave' default as False to allow us to reuse the function
so it can get both the full wave and a single packet wave.

When getting the wave I will read the pat_enable.. can be anything
the reason I do that is so that the more to follow will get set to 0.

If I do not do that there can be problems when calling like:

>>> cs.pat_wave
>>> cs.get_pat_wave()

when calling get_pat_wave().. It will return all packets after the first packet.